Turn the model picker's "More models" into an animated drill-down
Replace the Radix submenu / width-constrained-inline-expand split with a
single guided flow: "More models" and each provider now replace the
current dropdown content in place (root -> providers -> a provider's
models), with a back affordance at every step. Transitions between
steps slide + fade per the animations guidelines (short duration,
ease-out, reduced-motion aware).
